+++ tests/core/query/test_spatial_query.py
@@ -221,11 +221,7 @@ def test_query_raster(n_channels: int, is_labels: bool, is_3d: bool, is_bb_3d: b
     model = (
         Labels3DModel
         if is_labels and is_3d
-        else Labels2DModel
-        if is_labels
-        else Image3DModel
-        if is_3d
-        else Image2DModel
+        else Labels2DModel if is_labels else Image3DModel if is_3d else Image2DModel
     )
 
     image_element = model.parse(image)